Rework fixtures, add server fixture
