+++ tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .cots.marker import *
+from .cots.emergency import *
+from .server.fts_server import *
